refactor: Remove `ProxyIterator`.
