+++ src/Iterator/MultipleIterableIterator.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace loophp\collection\Iterator;
 
 use AppendIterator;
+use Iterator;
 use loophp\iterators\IterableIterator;
 use NoRewindIterator;
 
@@ -19,12 +20,17 @@ use NoRewindIterator;
  * @template TKey
  * @template T
  *
- * @extends ProxyIterator<TKey, T>
+ * @implements Iterator<TKey, T>
  */
-final class MultipleIterableIterator extends ProxyIterator
+final class MultipleIterableIterator implements Iterator
 {
     /**
-     * @param iterable<TKey, T> $iterables
+     * @var Iterator<TKey, T>
+     */
+    private Iterator $iterator;
+
+    /**
+     * @param iterable<TKey, T> ...$iterables
      */
     public function __construct(iterable ...$iterables)
     {
@@ -35,5 +41,36 @@ final class MultipleIterableIterator extends ProxyIterator
         }
 
         $this->iterator = $appendIterator;
+    }
+
+    /**
+     * @return T
+     */
+    public function current()
+    {
+        return $this->iterator->current();
+    }
+
+    /**
+     * @return TKey
+     */
+    public function key()
+    {
+        return $this->iterator->key();
+    }
+
+    public function next(): void
+    {
+        $this->iterator->next();
+    }
+
+    public function rewind(): void
+    {
+        $this->iterator->rewind();
+    }
+
+    public function valid(): bool
+    {
+        return $this->iterator->valid();
     }
 }
