+++ src/Iterator/ResourceIterator.php
@@ -11,6 +11,7 @@ namespace loophp\collection\Iterator;
 
 use Generator;
 use InvalidArgumentException;
+use Iterator;
 use loophp\iterators\ClosureIterator;
 
 use function is_resource;
@@ -18,13 +19,15 @@ use function is_resource;
 /**
  * @internal
  *
- * @template TKey
- * @template T
- *
- * @extends ProxyIterator<int, string>
+ * @implements Iterator<int, string>
  */
-final class ResourceIterator extends ProxyIterator
+final class ResourceIterator implements Iterator
 {
+    /**
+     * @var Iterator<int, string>
+     */
+    private Iterator $iterator;
+
     /**
      * @param false|resource $resource
      */
@@ -53,5 +56,36 @@ final class ResourceIterator extends ProxyIterator
             };
 
         $this->iterator = new ClosureIterator($callback, [$resource]);
+    }
+
+    /**
+     * @return string
+     */
+    public function current()
+    {
+        return $this->iterator->current();
+    }
+
+    /**
+     * @return int
+     */
+    public function key()
+    {
+        return $this->iterator->key();
+    }
+
+    public function next(): void
+    {
+        $this->iterator->next();
+    }
+
+    public function rewind(): void
+    {
+        $this->iterator->rewind();
+    }
+
+    public function valid(): bool
+    {
+        return $this->iterator->valid();
     }
 }
