+++ src/Iterator/TypedIterator.php
@@ -24,10 +24,15 @@ use function is_object;
  * @template TKey
  * @template T
  *
- * @extends ProxyIterator<TKey, T>
+ * @implements Iterator<TKey, T>
  */
-final class TypedIterator extends ProxyIterator
+final class TypedIterator implements Iterator
 {
+    /**
+     * @var Iterator<TKey, T>
+     */
+    private Iterator $iterator;
+
     /**
      * @param Iterator<TKey, T> $iterator
      * @param null|callable(mixed): string $getType
@@ -88,5 +93,36 @@ final class TypedIterator extends ProxyIterator
             },
             [$iterator]
         );
+    }
+
+    /**
+     * @return T
+     */
+    public function current()
+    {
+        return $this->iterator->current();
+    }
+
+    /**
+     * @return TKey
+     */
+    public function key()
+    {
+        return $this->iterator->key();
+    }
+
+    public function next(): void
+    {
+        $this->iterator->next();
+    }
+
+    public function rewind(): void
+    {
+        $this->iterator->rewind();
+    }
+
+    public function valid(): bool
+    {
+        return $this->iterator->valid();
     }
 }
